server/requirements.txt
fastapi
uvicorn[standard]
ultralytics
python-multipart
pillow
numpy
pydantic-settings
python-dotenv
